requirements.txt
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
click @ file:///D:/bld/click_1734858834134/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1733256735222/work
Flask @ file:///home/conda/feedstock_root/build_artifacts/flask_1741793020411/work
flask-cors==5.0.1
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1737420181517/work
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1733308265247/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1741263328855/work
MarkupSafe @ file:///D:/bld/markupsafe_1733219648915/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1733244044561/work
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1733160440960/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
